Add nice collision support

+++ src/GameComponent.h
@@ -6,7 +6,7 @@
 class GameComponent {
 public:
 	GameComponent();
-	virtual void Update(double) {}
+	virtual void Update(float) {}
 	virtual void FixedUpdate(float) {}
 
 	static std::vector<GameComponent*> components_array;
